// src/models/update-team-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GoMoney Football-League Manager Service
 * This service provides endpoints for all `football-league manager` related interactions
 *
 * OpenAPI spec version: 1.0.0
 * Contact: fatai@mail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * Update Team request data
 * @export
 * @interface UpdateTeamDto
 */
export interface UpdateTeamDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateTeamDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateTeamDto
     */
    code?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateTeamDto
     */
    logoUrl?: string;
}